pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "bioshop"
description = "bioinformatic shop tools"
requires-python = ">=3.7"
version = "0.0.1"
dependencies = [
    "requests",
    "edlib",
    "pysam",
    "xgboost",
    "scikit-learn",
    "pandas",
    "numpy",
    "portion",
    "fsspec",
    "humanize",
    "tqdm",
    "biopython",
    "pyfaidx",
    "cachier"
]

[tool.setuptools.packages.find]
include = ["bioshop*"]

[project.scripts]
newt = "bioshop.cmds.newt:main_cli"
